move card draw process
